# Remove the reserved member from rb_data_type_t as the addition of the compactor callback pushed it over a single cache line