@ioquatix 2019-08-20
Excellent investigation and good conclusion. Makes sense to me. The point of reserved field is to avoid breaking binary compatibility but I guess that happened anyway.