@methodmissing 2019-08-21
The Ubuntu CI pipelines are having a hard time and failing with:

```
E: Failed to fetch https://packages.microsoft.com/repos/microsoft-ubuntu-bionic-prod/dists/bionic/main/binary-amd64/Packages.bz2  File has unexpected size (66537 != 66619). Mirror sync in progress? [IP: 40.76.35.62 443]
    Hashes of expected file:
    - Filesize:66619 [weak]
    - SHA512:4754fea65f57f35fc52e332a6d5cddaf65ea00bbd9e9f428d3a4325c1ac26856190ae82022cdb7a41bf1337e5c7fb1a429933bae9817bacaa8a5186c2aae9f1b
    - SHA256:f971467de90380c9920956be80642d7f6eb0e75042822ea4537d9f5190e13d9f
    - SHA1:0f8e0e8ac649a20ef2029c9f4ffc6bc1ef559257 [weak]
    - MD5Sum:8aa7d0d41b74ea8b84a434a726ea92fa [weak]
   Release file created at: Tue, 20 Aug 2019 22:43:21 +0000
E: Some index files failed to download. They have been ignored, or old ones used instead.
##[error]Process completed with exit code 100.
```

Will rebase later today with master to trigger a new build as I don't seem to be able to trigger manually just for those.