@methodmissing 2019-09-22
@ioquatix @nobu left some feedback in https://bugs.ruby-lang.org/issues/16125#note-1 (`dcompact` consumed an element of `reserved`, `pahole` didn't properly report that case).

I still think the `reserved` member should now be removed in declaration as it's now being consumed by `dcompact` anyways. For consistency. Thoughts?