@ioquatix 2019-09-22
If it doesn’t affect binary compatibility and it improves performance, it makes sense to me.